+++ cardano-api/cardano-api.cabal
@@ -404,6 +404,7 @@ test-suite cardano-api-test
     Test.Cardano.Api.EpochLeadership
     Test.Cardano.Api.Eras
     Test.Cardano.Api.Experimental
+    Test.Cardano.Api.Experimental.Fee
     Test.Cardano.Api.Genesis
     Test.Cardano.Api.GovAnchorValidation
     Test.Cardano.Api.IO
